+++ src/releases/archived.rs
@@ -3,7 +3,7 @@ use crate::{
     helpers::{get_document, get_file_stem},
     package::{Build, Os, Package},
     releases::ReleaseType,
-    settings::CONFIG_PATH,
+    settings::SETTINGS,
 };
 use async_trait::async_trait;
 use chrono::NaiveDateTime;
@@ -211,6 +211,6 @@ impl ReleaseType for Archived {
     }
 
     fn get_db_path(&self) -> PathBuf {
-        CONFIG_PATH.parent().unwrap().join("archived_db.bin")
+        SETTINGS.read().unwrap().databases_dir.join("archived.bin")
     }
 }