+++ src/releases/daily.rs
@@ -3,7 +3,7 @@ use crate::{
     helpers::{get_document, get_file_stem},
     package::{Build, Os, Package},
     releases::ReleaseType,
-    settings::CONFIG_PATH,
+    settings::SETTINGS,
 };
 use async_trait::async_trait;
 use chrono::{Datelike, NaiveDateTime, Utc};
@@ -97,6 +97,6 @@ impl ReleaseType for Daily {
     }
 
     fn get_db_path(&self) -> PathBuf {
-        CONFIG_PATH.parent().unwrap().join("daily_db.bin")
+        SETTINGS.read().unwrap().databases_dir.join("daily.bin")
     }
 }